More prototype updates (nwlib, wmake)
[git-p4: depot-paths = "//depot/openwatcom/": change = 24631]

+++ bld/nwlib/c/orlrtns.c
@@ -35,7 +35,7 @@
 static orl_handle       ORLHnd;
 static orl_funcs        OrlFuncs;
 
-static void * ObjRead( void *fil, int len )
+static void * ObjRead( void *fil, size_t len )
 /*****************************************/
 {
     buf_list *  buf;
